Add delete confirmation dialog box
Close #41

+++ resources/js/Page/CompanyProfile.jsx
@@ -7,8 +7,11 @@ import StarIcon from "@mui/icons-material/Star";
 import Box from "@mui/material/Box";
 import DeleteIcon from "@mui/icons-material/Delete";
 import AddCircleIcon from "@mui/icons-material/AddCircle";
+import ConfirmDialog from "../components/Confirm.jsx";
 
 export default function CompanyProfile(props) {
+    const [deleteOpen, setDeleteOpen] = useState(false);
+
     return (
         <div>
             <Grid container justifyContent="space-between">
@@ -32,25 +35,28 @@ export default function CompanyProfile(props) {
                     >
                         Sign Out
                     </Button>
+
                     <Button
                         sx={{ backgroundColor: "lightgrey" }}
                         variant="contained"
                         startIcon={<DeleteIcon />}
-                        onClick={() => {
+                        onClick={() => setDeleteOpen(true)}
+                    />
+                    <ConfirmDialog
+                        title="Delete account ?"
+                        open={deleteOpen}
+                        setOpen={setDeleteOpen}
+                        onConfirm={() => {
                             fetch("/api/companies/" + props.data.company.id, {
                                 method: "delete",
-                                headers: {
-                                    "X-CSRF-TOKEN": props.csrf_token,
-                                },
-                                body: {
-                                    _token: props.csrf_token,
-                                },
                             }).then((res) => {
-                                if (!props.data.user.account_id)
+                                if (!props.data.company.account_id)
                                     window.location.replace("/signout");
                             });
                         }}
-                    ></Button>
+                    >
+                        Are you sure you want to delete this account?
+                    </ConfirmDialog>
                 </div>
             </Grid>
             <Grid container sx={{ mt: 3, justifyContent: "center" }}>
